Fixed off-position node problem in CP layout
