+++ src/features/HierarchyViewer/utils/hierarchy-util.ts
@@ -1,8 +1,13 @@
+import { HierarchyNode } from 'd3-hierarchy'
 import { Visibility } from '../../../models/CxModel/NetworkSummary/Visibility'
 import { NdexNetworkSummary } from '../../../models/NetworkSummaryModel'
-import { ValueType } from '../../../models/TableModel'
+import { Table, ValueType } from '../../../models/TableModel'
+import { D3TreeNode } from '../components/CustomLayout/D3TreeNode'
 import { HcxMetaData } from '../model/HcxMetaData'
 import { HcxMetaTag } from '../model/HcxMetaTag'
+import { IdType } from '../../../models/IdType'
+import { NodeView } from '../../../models/ViewModel'
+import { CirclePackingView } from '../model/CirclePackingView'
 
 export const getHcxProps = (
   summaryObject: Record<string, any>,
@@ -84,4 +89,118 @@ export const isHCX = (summary: NdexNetworkSummary): boolean => {
   const metadata: HcxMetaData | undefined = getHcxProps(networkPropObj)
 
   return metadata !== undefined
+}
+
+const findAllDescendants = (
+  root: HierarchyNode<D3TreeNode>,
+  targetNodeId: string,
+): string[] => {
+  let queue: HierarchyNode<D3TreeNode>[] = [root]
+  // Find the target node
+
+  let targetNode: HierarchyNode<D3TreeNode> | undefined
+  while (queue.length > 0) {
+    const node = queue.shift()
+    if (node === undefined) {
+      continue
+    }
+    if (node.data.id === targetNodeId) {
+      targetNode = node
+      break
+    }
+    if (node.children) {
+      const children = node.children as HierarchyNode<D3TreeNode>[]
+      queue.push(...children)
+    }
+  }
+  // Collect all descendants' IDs
+  queue = [targetNode as HierarchyNode<D3TreeNode>]
+  const descendants: string[] = []
+  while (queue.length > 0) {
+    const node = queue.shift()
+    if (node === undefined) {
+      continue
+    }
+    descendants.push(node.data.id)
+    if (node.children) {
+      const children = node.children as HierarchyNode<D3TreeNode>[]
+      queue.push(...children)
+    }
+  }
+  return descendants
+}
+
+const SCALING_FACTOR = 40
+
+export const applyCpLayout = (
+  cpViewModel: CirclePackingView,
+  subsystemNodeId: string,
+  interactionNetworkId: IdType,
+  interactionNetworkTable: Table,
+  nodeViews: Record<string, NodeView>,
+): Map<IdType, [number, number]> => {
+  // ID format is parentID-node name
+  // const cpViewModel = getCpViewModel()
+  if (cpViewModel === undefined) {
+    return new Map()
+  }
+
+  const { hierarchy } = cpViewModel
+  if (hierarchy === undefined) {
+    return new Map()
+  }
+
+  const allDescendants: string[] = findAllDescendants(
+    hierarchy,
+    subsystemNodeId,
+  )
+
+  // Filter to get only the leaf nodes
+  const leavesInCircle = allDescendants.filter(
+    (nodeId: string) => nodeId.split('-').length > 1,
+  )
+
+  const cpNodeViews = cpViewModel.nodeViews
+  const { rows } = interactionNetworkTable
+
+  const id2name: Map<IdType, string> = new Map()
+  Object.keys(nodeViews).forEach((nodeId: string) => {
+    const row = rows.get(nodeId)
+    if (row === undefined) {
+      return
+    }
+    const nodeName: string = row.name as string
+
+    // Find the name with prefix
+    leavesInCircle.find((cpNodeId: string) => {
+      const parts: string[] = cpNodeId.split('-')
+      if (parts[1] === nodeName || `${parts[1]}-${parts[2]}` === nodeName) {
+        id2name.set(nodeId, cpNodeId)
+      } else {
+        const pattern = /^-?\d+$/
+        if (pattern.test(parts[1])) {
+          if (parts[2] === nodeName || `${parts[2]}-${parts[3]}` === nodeName) {
+            id2name.set(nodeId, cpNodeId)
+          }
+        }
+      }
+    })
+
+    // id2name.set(nodeId, prefix + nodeName)
+  })
+
+  const positionMap: Map<IdType, [number, number]> = new Map()
+  id2name.forEach((nodeName: string, nodeId: string) => {
+    const cpNodeView = cpNodeViews[nodeName]
+    if (cpNodeView === undefined) {
+      return
+    } else {
+      positionMap.set(nodeId, [
+        cpNodeView.x * SCALING_FACTOR,
+        cpNodeView.y * SCALING_FACTOR,
+      ])
+    }
+  })
+
+  return positionMap
 }
